Removed Models Folder/Class - Switching to Singular Class over Model
I decided I am going to KISS and do a single class which writes to the database. I was originally planning on utilizing a system which I had seen in a production environment with data validation and classes being seperate. That seems better for a bigger project, but uneeded here.